added authorization on routes

+++ routers/posts.py
@@ -1,7 +1,7 @@
 from typing import Annotated
 from sqlalchemy.ext.asyncio import AsyncSession
 from database import get_db
-
+from auth import  Current_user
 from fastapi import APIRouter, Depends, HTTPException, status
 from sqlalchemy import select
 from sqlalchemy.orm import selectinload
@@ -19,30 +19,31 @@ async def get_post(post_id:int,db:Annotated[AsyncSession, Depends(get_db)]):
     return post
 
 @router.put("/{post_id}",response_model=PostResponse)
-async def update_full_post(post_id:int, post_data: PostCreate, db:Annotated[AsyncSession, Depends(get_db)]):
+async def update_full_post(post_id:int, post_data: PostCreate,current_user: Current_user,db:Annotated[AsyncSession, Depends(get_db)]):
     result=await db.execute(select(models.Post).where(models.Post.id==post_id))
     post=result.scalars().first()
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found")
 
-    if post.user_id != post_data.user_id:
-        result=await db.execute(select(models.User).where(models.User.id==post_data.user_id))
-        user=result.scalars().first()
-        if not user:
-            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="User not found")
+    if post.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to update this post")
     post.title=post_data.title
     post.content=post_data.content
-    post.user_id=post_data.user_id
+    
     await db.commit()
     await db.refresh(post, attribute_names=["author"])  # refresh the post to get the author relationship loaded, we can also use options(SelectinLoad(models.Post.author)) when querying the post to load the author relationship
     return post
 
 @router.patch("/{post_id}",response_model=PostResponse)
-async def update_post_partial(post_id:int, post_data: PostUpdate, db:Annotated[AsyncSession, Depends(get_db)]):
+async def update_post_partial(post_id:int, post_data: PostUpdate, current_user: Current_user, db:Annotated[AsyncSession, Depends(get_db)]):
     result=await db.execute(select(models.Post).where(models.Post.id==post_id))
     post=result.scalars().first()
     if not post:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found")   
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found") 
+
+    if post.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to update this post")
+ 
 
     update_data=post_data.model_dump(exclude_unset=True)
     for key, value in update_data.items():
@@ -57,14 +58,11 @@ async def update_post_partial(post_id:int, post_data: PostUpdate, db:Annotated[A
 
 
 @router.post("",response_model=PostResponse,status_code=status.HTTP_201_CREATED)
-async def create_post(post: PostCreate,db :Annotated[AsyncSession, Depends(get_db)]):
-    result=await db.execute(select(models.User).where(models.User.id==post.user_id))
-    user=result.scalars().first()
-    if not user:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="User not found")
+async def create_post(post: PostCreate,current_user: Current_user,db :Annotated[AsyncSession, Depends(get_db)]):
+    
 
     new_post=models.Post(
-        user_id=post.user_id,
+        user_id=current_user.id,
         title=post.title,
         content=post.content
     )
@@ -81,10 +79,12 @@ async def get_posts(db:Annotated[AsyncSession, Depends(get_db)]):
 
 
 @router.delete("/{post_id}",status_code=status.HTTP_204_NO_CONTENT)
-async def delete_post(post_id:int,db:Annotated[AsyncSession, Depends(get_db)]):
+async def delete_post(post_id:int,current_user: Current_user,db:Annotated[AsyncSession, Depends(get_db)]):
     result=await  db.execute(select(models.Post).where(models.Post.id==post_id))
     post=result.scalars().first()
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found")
+    if post.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to delete this post")
     await db.delete(post)
     await db.commit()
